requirements.txt
# pip install -r ./requirements.txt
Django==2.2.1
djangorestframework
django-cors-headers
whitenoise
dj-database-url
gunicorn
psycopg2
pytz
django-ckeditor
boto3
django-storages
python-slugify